Extracted getVersesList to separate file
