+++ src/Pages/SongPage/PresentationMode/PresentationMode.tsx
@@ -8,6 +8,7 @@ import "@mantine/carousel/styles.css";
 
 import classes from "./PresentationMode.module.scss";
 import { Hymn } from "../../../utils/types";
+import { getVersesList } from "./getVersesList";
 
 export function PresentationMode({
     presenting,
@@ -37,47 +38,7 @@ export function PresentationMode({
     }, [fullscreen, presenting, toggle]);
 
     useEffect(() => {
-        const [type, content] = selectedHymn.content
-            ? ["html", selectedHymn.content]
-            : selectedHymn.markdown
-            ? ["markdown", selectedHymn.markdown]
-            : [null, ""];
-
-        if (type) {
-            const formattedVerses =
-                type === "html"
-                    ? content
-                          .split(
-                              // Line breaks and end of headings or paragraphs
-                              /(<br\s*\/?>\s*<br\s*\/?>)|(<\/h\d>)|(<\/p>)|(<\/b>(<br\s*\/?>)?<p>)/g
-                          )
-                          .map((item) => item?.replaceAll(/<br\s*\/?>/g, "\n"))
-                          .map((item) =>
-                              // Strip out HTML tags
-                              item
-                                  ?.replaceAll(/<\/?(\w|\d|\s|#|"|'|=)*>/g, " ")
-                                  .trim()
-                          )
-                          .filter((item) => item && item.trim() !== "")
-                    : content.split("\n\n");
-
-            const finalVerses: Array<string> = [];
-            const chorus = formattedVerses.find((element) =>
-                /CHORUS/i.test(element)
-            );
-
-            // Based on: https://github.com/TinasheMzondiwa/cis-android/blob/33ad4066f409215780b11ff813e5fb77ee806c46/app/src/main/kotlin/com/tinashe/hymnal/ui/hymns/sing/present/PresentPagerAdapter.kt#L20
-            for (const [index, verse] of formattedVerses.entries()) {
-                finalVerses.push(verse);
-                if (index > 2 && chorus && !/CHORUS/i.test(verse)) {
-                    finalVerses.push(chorus);
-                }
-            }
-
-            setVerses(finalVerses);
-        } else {
-            setVerses([]);
-        }
+        setVerses(getVersesList(selectedHymn));
     }, [selectedHymn, setVerses]);
 
     return (
